pyproject.toml
[project]
name = "rl-youtube"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dnspython==2.6.1",
    "google-api-python-client==2.131.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",
    "gymnasium==0.29.1",
    "numpy==1.26.4",
    "pydantic-settings>=2.13.1",
    "pymongo[srv]==4.7.2",
    "python-dateutil==2.9.0",
    "tensorflow-cpu>=2.16.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
